refactor: でんしょうおの大気レイヤーを分離
